+++ apps/vscode/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.118.0 (unreleased)
+## 1.118.0 (Release on 2024-11-26)
 
 - Provide F1 help at cursor in Positron (<https://github.com/quarto-dev/quarto/pull/599>)
 - Expose new context keys for the main language of a document (<https://github.com/quarto-dev/quarto/pull/608>)